requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
docker-pycreds==0.4.0
filelock==3.6.0
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.4.0
idna==3.3
joblib==1.1.0
numpy==1.22.3
packaging==21.3
pathtools==0.1.2
pip==21.2.4
promise==2.3
protobuf==3.20.0
psutil==5.9.0
pyDeprecate==0.3.2
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
sentry-sdk==1.5.10
setproctitle==1.2.3
setuptools==58.0.4
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
tokenizers==0.11.6
torch==1.11.0
torchmetrics==0.7.3
tqdm==4.64.0
transformers==4.17.0
typing_extensions==4.1.1
urllib3==1.26.9
wandb==0.12.14
wheel==0.37.1
